remove optional overhead
You must use option in two cases:
- a method call you do returns an Optional
- you return an Option so someone else has to check nullability

You must not create an Optional to immediately check nullability,
it kills the purpose of the Optional. The purpose of the Optional is to
force people to check null values to avoid NullPointerException. There
you are forcing yourself to check null values. You can just check null
value right now and it's over.